requirements.txt
pytz==2014.1
mozdef_lib
duo_client
